Move FindUsedBlobs to package restic
